Removed /index from the imports and refactored the Search component
