+++ src/components/AllBooks/index.js
@@ -1,7 +1,5 @@
 import { Link } from "react-router-dom";
-import { Container } from "./allBooks.styles";
-import { Card } from "./allBooks.styles";
-import { Genre } from "./allBooks.styles";
+import { Container, Card, Genre } from "./allBooks.styles";
 
 export const AllBooks = (props) => {
   const mappedBooks = props.data.map((book) => {
